Implement Checks on Revisions
This commit adds Checks as an entity that Checkers can report on patch
sets. It adds logic to store the entity in NoteDb, a Java API, a REST
API, a test API and very basic test coverage.

Checks are stored in Git Notes keyed by the SHA1 of the patch set they
belong to. NotesMap is used to read from and write to NoteDb.

This change leaves a couple of TODOs for future changes:
1) The test coverage is *very* basic, but the test framework is there,
   so this work can easily be sharded.
2) The JavaDoc needs some love.
3) As a performance improvement, we want to write notes individually
   instead of in bulk. The same applies to reading.
4) Smaller implementation nits here and there where we throw an
   UnsupportedOperationException or left a TODO.
5) A cache for Checks and Check status.
6) Checking that the provided checkerUUIDs match actualy Checkers
7) Permissions to prevent Checkers from updating other Checkers Checks'

Change-Id: I5109764e2ba776c6e1dcc630559b2ef1a40434ac