Support externally-provided UUIDs with a scheme: prefix
UUID is now provided by callers, as opaque URIs of the form SCHEME ':'
ID. The scheme is analogous to the scheme in an ExternalId, and
describes the system from which the checker originated. The ID portion
is a unique ID in any format, as long as it is URL-encoded.

Similar to the "id" field in ChangeInfo, the "uuid" field in CheckerInfo
can be passed verbatim in a /plugins/checkers/X URL without
double-encoding. This leads to some complexity in the resource parsing
code, but the upside is it's consistent with changes and we never have
to deal with double-encoded URLs.

We still use SHA-1s as the lookup key for ref names, it's now just
SHA1(UUID). This leads to some complexity in the CheckerConfig class,
since an instance may be created either with a CheckerUuid or a ref
name. In the latter case, the UUID is not known until the config is
loaded.

When parsing a UUID, we make sure to normalize the %-encoding and scheme
case, so that the value we're taking the SHA-1 of and using as a key in
a map is consistent.

Name is now a completely optional field, treated like any other optional
field. In particular, we no longer have special logic for recording
renames in commit messages. The logic would have had to get more
complicated to take into account unset names, and I judged it not worth
the effort. If we go down that route, perhaps we can write something
more general for recording updates to other fields (again: name is not
special), but honestly I think `git log -p` is good enough.

Change-Id: I0fff95b3f9eb408cee4651544b0b5acbef30e016