WIP/RFC: Include implicit relevant checkers in getChecks
Checkers that are relevant for the change should have their checks
reported to users even if the external system has not yet reported any
check status. Keep track of which checkers associated with the repo are
not found in the check map, and backfill the results with synthetic
entries for any relevant ones.

Change-Id: I2d105efecb83c0c6c29fa0c8a6eee323d39bdb2f