Consistently return all checks from NoteDb for Get Check and List Checks
In our recent discussions we decided that for the Get Check and List
Checks REST endpoints we always want to return all checks that exist in
NoteDb, regardless whether the corresponding checker is diabled, no
longer applying (repo or query changed), not-existing or invalid (not
parseable).

Reasons for returning those checks:
1. better UX as checks that have been once shown on the screen do not
   magically disappear because the checker no loger applies
2. the UI has full control over how to visiulize these checks (the UI
   knows the checks and the states of the corresponding checkers)

Before this change the behaviour was not consistent, e.g. it was
possible to get checks of a checker that no longer applies because the
query was changed, but it was not possible to get checks of a checker
that no longer applies because the repository was changed.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I523254403375d87af6f7c23c21e72417e39e0bd7