Update patch set 3
Patch Set 3:

I decided that the bakfilling should happen in ListChecks, not in NoteDbChecks. The NoteDbChecks implementation should only be returning what is stored in NoteDb; it shouldn't need to pull in a dependency on the query subsystem to do so.

Patch-set: 3
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>