Ensure check data is deleted when change is deleted
We are lucky that when a change is deleted, all refs starting with
refs/changes/<sharded-change-id> are removed. Since we use that prefix
for the refs of the checks, we don't have to implement anything
additionally.

To ensure that unrelated changes in core Gerrit do not silently break
the check data deletion, add a test for it.

The existing Java API for checks didn't behave well when the change was
deleted. Correct that and also test it.

Change-Id: I40187d964a302bfd0112c7f02f698d0b7d51fb67