Move "CheckBackfiller" out from the "db" package
This class doesn't necessarily need to be part of
the "api" package because ideally we would like
to keep the "api" package small and mainly contains
"api" facing classes.

Moving this class to the "db" package to make it
an class for internal implementation.

This commit has a side effect because it makes
the class and its two methods public. In the follow
up commit, we will move the existing backfill code
from those APIs, e.g. ListChecks, to "NoteDbChecks".
At the meantime, we will extead "Checks" API to
allow callers to specify an option which controls
whether "backfill" should be done in "GetCheck"
or "GetChecks". After that, we could mark
"CheckBackFiller" as package-private again.

Change-Id: I0859fa50e7f666fc83a6b0adf0742ff681c9bdb3