Add #getCombinedCheckState to the Check interface
This method is useful for callers who want to know the
combined check state of a patch set. For example,
in the follow-up change, the "ChecksSubmitRule" would
simply call this API to get the combined check state
of the latest patch set and then use that state to
decide whether a change is submittable or not.

Basic tests are added to verify the bahavior of this
API, more specific behaviors are covered by existing
"CombinedCheckStateTest".

Change-Id: I341dfdab8beaccd28d567d31196ab9a542661d24