Add tests with invalid checker queries
Checker queries can be invalid in 2 ways:
1. the query is not parseable
2. the query contains non-allowed predicates

Creating checkers with invalid queries is not allowed, however if it
happens the system should still behave properly, hence this should be
tested.

Checkers with invalid queries should be treated the same way as checkers
with valid, non-matching queries. That's already the case but this
change adds tests for this.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I8baca2b117c3196272a0c592ce71cd9a4b733754