Add class with checker test data
Many test classes have tests with invalid data for checkers. Instead of
defining the invalid data again and again in each class define it once
in a new CheckerTestData class. Having this defined in a common place
makes the tests consistent and allows us to tests our assumptions about
this test data.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I5708fb80adbd5c6aa72dd10c4ccf0c5ca1f6e049