Update patch set 1
Patch Set 1: Code-Review+2

I'm not totally convinced that we want to add a source-level dependency on ChangeData + ChangeQueryBuilder to this class which was previously a simple value type. But I'm fine with it for now, we can always move it back if we have a compelling reason.

Patch-set: 1
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+2