CheckerConfig: Set new updated timestamp only when config has changed
If there was a no-op update CheckerConfig didn't create a new commit but
it was still setting a new updated timestamp on the loadedChecker. As
result of this the Update Checker REST endpoint returned a wrong updated
timestamp, making the impression that a new commit was created.

Add a test to verify that the updated timestamp is correctly returned
for no-op updates. In addition also add a test to verify that no-op
updates don't create a new commit (but this test would have succeeded
even before this change).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Iadc96013a8993369775e29d67c26a8965e294d02