Consistently use the name 'repository' in the checks API
Some parts of the checks API used the term 'repository' (CheckerInfo)
while other parts of the checks API used the term 'project' (CheckInfo,
CheckablePatchSetInfo). The checks API should be consistent with the
naming.

Reasons for using the term 'repository':
- this is the term that is used in the UI
- the checks API will likely be used by people that are not familiar
  with Gerrit yet (and hence are not familiar the 'project' term), e.g.
  owners of analyzers that want to integrate with Gerrit
- it's the same name that is used by other git services

Reasons for using the term 'project':
- it's consistent with the Gerrit core REST API
- it's the name that is familiar for existing Gerrit users

I discussed this with aliceks@ and we think that 'repository' is the
better choice. Consistency with the Gerrit core REST API is not that
important since the checks API is rather independent from it.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I0e1d68277e4829f270e5f247673757eff9990fea