ListChecksIT: Let each test care about the test setup itself
At the moment ListChecksIT creates some checkers in its setUp() method.
This makes the tests less readable since for understanding the
assertions in one test one always has to take the global test setup into
account (e.g. one needs to scroll to the beginning of the class to see
the global setup).

It also makes the test assertions more complicated since this way many
tests do also check conditions that are not relevant for them. Not
needing to do this simplifies the tests.

Also some pieces of the global test setup were only relevant for some of
the tests (e.g. setting a checker name, setting check states). So it was
getting more and more complicated to have a common global test setup
that suited all tests.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I6615ecf83f3671db5e90a45cfd3d46c6e978e43a