GetCheckIT: Only assert things that are relevant to the tests
Many of the tests asserted things that were not relevant to them (e.g.
they verified unrelated fields in CheckInfo). This made the tests more
prone to be flaky, hard to read and difficult to maintain.

Add a new getCheck() test so that there is at least one test that checks
the full returned CheckInfo.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I70de219dd2b2e640e1276d656bb33ee2a58eedd2