Update patch set 2
Patch Set 2: Code-Review-1

I don't understand. Why is this operator anything other than a no-op? How could "query pending checks" possibly return checks that are *not* pending? That's a logical contradiction.

If there is a real use case we want to support, then IMHO we need to fix the naming of either the endpoint or the operator to avoid this contradiction. I don't want to be in a position of explaining in the documentation that "query pending checks" returns pending checks except if you do/don't include some operator in which case it might include checks that are pending according to definition A but not pending according to definition B.

Patch-set: 2
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1