Update patch set 1
Patch Set 1:

I'm leaving for vacation momentarily so I can't review these soon; passing on to Alice.

That said, I still have the same concern that I had on the previous implementation: I don't want to use different logic for the submit rule than we are exposing in the combined check state chip in the UI. Can we adapt the combined check state chip to match the logic in the submit rule?

Patch-set: 1
CC: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>