Test anonymous calls to REST endpoints and make the behavior consistent
In Gerrit core REST endpoints that require authentication reject
anonymous requests with "Authentication required". Do this for the check
REST endpoints too to be consistent. This message is nice for users
which have the needed permissions but just forget to authenticate (e.g.
that forget the '/a' prefix in the URL).

Checks can be accessed anonymously if the change is visible to anonymous
users.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ic88f05530091e25b63a4a925488857da46109a87