Update patch set 8
Patch Set 8:

Ping. It would be nice if we could submit this soon. Hanwen is starting to implement a test checker and it would be nice if he doesn't need to use the old API, just to be broken when this change gets submitted.

Patch-set: 8
Reviewer: Gerrit User 1003935 <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>