Update patch set 2
Patch Set 2: Code-Review-1

Something smells wrong that you have to repeat this check in all REST API endpoints. There are only 11 endpoints in all of the restapi packages in core that throw this exception. Why is it not necessary for the majority of endpoints in core?

Patch-set: 2
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1