Query pending checks: Make is:<STATE> operators consistent
Some operators require an underscore in the state name (e.g.
'NOT_STARTED', 'NOT_RELEVANT'), while others don't accept it
('INPROGRESS'). This is because the former are mapped to the CheckState
enum, and the latter isn't. From a users perspective this difference is
not obvious. This is why we now support both forms, with and without
underscore, in all cases. This means 'NOTSTARTED', 'NOTRELEVANT' and
'IN_PROGRESS' are working now too.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ic6352306cd196751af9d6a2728433d312eb5f361