Handle invalid data in check json properly
NoteDbCheck is the representation of a Check that can be serialized to
JSON. Fields in NoteDbCheck were typed and as result deserializing JSON
back to NoteDbCheck could fail if these fields contained invalid data
(e.g. if the 'created' field was set to a string that couldn't be parsed
as a Timestamp). Other invalid data (e.g. invalid state) was silently
ignored. Another category of problems that were not handled are required
fields (created and updated) that are not set in the JSON.

To properly handle these cases we read all fields as strings and do the
parsing into types explicitly in NoteDbCheck on conversion to a Check
instance. This allows us to handle all error conditions and throw
ConfigInvalidException with a descriptive message.

When listing checks invalid checks are ignored, but if the checker is
relevant backfilling is done.

Follow-up changes should make sure that non-parseable data (invalid
JSON) is handled properly.

Change-Id: If69584abad4e79ee284d3d537eca25efc6e0b11a
Signed-off-by: Edwin Kempin <ekempin@google.com>