Update patch set 1
Patch Set 1:

What's the motivation here? ISTM we have the exact same problem with JSON comments in NoteDb in core, but we've never had any corruption in the 2 years or so that we've been using that format. I mean, more safety is better than less safety, but this seems like a lot of work to get there.

I think what you might be saying is that you don't like the inherent unsafety of the default Gson behavior of populating mutable, nullable fields. If that's the root of the issue, then maybe we should explore JSON parsing options other than bolting a lot of postcondition checks onto the existing broken model.

Random ideas:

* Use something like https://stackoverflow.com/questions/21626690/gson-optional-and-required-fields to mark certain fields as required.
* Consider writing a type adapter for AutoValue types.

I guess the options I just described would result in exceptions (likely unchecked exceptions) when there's invalid data, which suggests the answer to the question of what to do with invalid checks is to fail the whole operation. Would that be so bad? If corruption were common, yeah, then a small mishap could result in bringing CI to a halt, but see above: our experience with comments suggests that corruption is essentially nonexistent.

Patch-set: 1
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>