Update patch set 1
Patch Set 1:

The motivation is to have a proper error message when invalid data occurs so that we can easily find and fix it.

I agree that invalid data is rather unlikely and it seems it's not worth the effort to have the most possible safety here. When I started this I thought, it could be easily done, but as it turned out to be more and more work, I thought I better ask for opinions by pushing this as RFC.

What you say makes sense. Your suggestion to write a type adapter looks promising. I will have a closer look at this when there is time and abandon this change for now.

Patch-set: 1
Reviewer: Gerrit User 1003935 <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>