Update patch set 1
Patch Set 1:

(4 comments)

> Patch Set 1:
> 
> I'm leaving for vacation momentarily so I can't review these soon; passing on to Alice.
> 
> That said, I still have the same concern that I had on the previous implementation: I don't want to use different logic for the submit rule than we are exposing in the combined check state chip in the UI. Can we adapt the combined check state chip to match the logic in the submit rule?

I understand your concern. On the other hand, tying the combined check state and the submit rule completely together also introduces some issues and might not always be better for users. For instance, if we decided to keep only one 'in progress' state, we would unnecessarily block users even for optional checks. If we don't want that, we would need two different 'in progress' states, which in turn would also confuse users or could involve very lengthy UI explanations.

From a long-term perspective, tying the combined check state and submittability together would also leave us less flexibility when adding more individual check states. Of course, this always includes the risk that we forget a case or introduce a strange behavior but I'm willing to take it. Adding a second 'in progress' state in the future (if we change our opinion) will be easier than deprecating such a state. Hence, we'll try this different approach for now.

By the way, we hope to address the concern you mentioned by adding additional details to the combined check state on the UI. We could say "2 of 3 in progress (1 required)". It's something we intend to do also for the 'failure' state to make the distinction with 'warning' more obvious. Thus, we rather encourage users to watch out for required checks and what their state is instead of encrypting such details in the combined check state.

Patch-set: 1
Reviewer: Gerrit User 1030912 <1030912@173816e5-2b9a-37c3-8a2e-48639d4f1153>