Checks test API: Allow easier specification of required/optional
Using blocking conditions instead of a required flag is a mechanism
which will allow us quite some flexibility in the future. However,
most tests don't need that flexibility. They only need the notion
of a required or optional checker - even if we extend the blocking
conditions in the future. Hence, introduce those much easier concepts
to the test API. In addition, adjust tests where reasonable to this
different notion. This should also keep those tests more stable in the
future.

Change-Id: Ic443771994931aa4596f11e7839d3a1a33ad27b3