Don't support checks for change edits
Change edits are only visible to change owners
and thus are not useful to checker systems. Thus
this commit adds additional checks to reject
if the caller tries to create/get/list checks
on a change edit.

Instead of doing like this commit, there is another
option which simply returns an empty check set
for a change edit. However, this opens the door
for the client to depend on an unuseful feature
which may break them if we decide to turn it off
in the future.

Change-Id: Icdcbca3afb070813f6cb829c1de787eff1c63782