Add backend support for copying checks across patchsets.
When a new patchset would be created, normally, all checks needed to
rerun. With this added backend support, users could specify when check
results should be copied to the next patchset.

We will copy checks whenever the copyPolicy contains the ChangeKind of
the current revision. E.g if the ChangeKind is 'NO_CHANGE' and the
copyPolicy for a checker is {'NO_CHANGE', 'NO_CODE_CHANGE'} we would
copy the check. The set of copyPolicy may also be empty and then
checks would never be copied. Also, the check needs to be finished
(SUCCESSFUL, FAILED, or NOT_RELEVANT).

The way we copy checks is by listening to each new revision created
using onRevisionCreated. For each revision, we will loop over all of the
checks and for each check that satisfies the conditions above, we will
copy it completely (message, url, state, started, finished). By building a
new check for the current revision and updating it with all of the
details of the check that is being copied.

The advantage of this approach is that it allows flexibility for the
users, as they can specify exactly which changes should carry-over
previous checks.

The disadvantage would be the fact that users have to specify exactly
when they want to carry over checks for each new checker they add.

The alternative was to decide ourselves when to copy checks (e.g always
copy when there is no change) but it is better to let the users decide
what they would like to do.

Change-Id: Id2b0115dc7cb6090ca2e9a88029683d67247a2a1