Send email notification if combined check state is updated
When the combined check state of a patch set changes by
creating/updating a check we now send an email notification to the
change owner, the reviewer, users who starred the change and project
watchers.

We only send a notification when the combined check state changes
because sending a notification each time a check is created or updated
would result in too many emails and we don't want to spam users.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ibb5cd800b786f6754fcf4f3f28f26c2cad7ded94