Update patch set 1
Patch Set 1:

Is it possible to add this functionality partially in core?  I.e. we already have an interface to allow plugins to provide templates for mail sending, so can that be extended to also do the extraction of examples?  Otherwise it looks like all plugins that have templates will need to repeat the same code.

Patch-set: 1
CC: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>