Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Is it possible to add this functionality partially in core?  I.e. we already have an interface to allow plugins to provide templates for mail sending, so can that be extended to also do the extraction of examples?  Otherwise it looks like all plugins that have templates will need to repeat the same code.

Gerrit core extracts the templates during site initialization. At that moment plugins are not loaded yet so the site intializer cannot ask the plugins for their templates.

What would be possible is to make the extractMailExample method from Gerrit core callable for plugins, but not sure if this saves much.

Patch-set: 1